+++ src/routes/pages/Home.tsx
@@ -1,7 +1,7 @@
 import Category from '@/components/common/Category'
 import Feed from '@/components/playlist/Feed'
 import { useHeaderStore } from '@/stores/header'
-import { useFetchPlaylists } from '@/hooks/playlists'
+import { useFetchPlaylists } from '@/hooks/useFetchPlaylists'
 import { useEffect } from 'react'
 
 export default function Home() {
